msm8996-common: update TWRP to oreo device tree (TeamWin)
If we ever want to have our own TWRP builds with our minimal Kernel
(recovery defconfig) building inline with LineageOS 15.1.

Some blobs were lazy-imported from TeamWin just to hold them in place
over here, we could opt-in for a vendor copy of those blobs too if we
weren't that lazy.

Change-Id: I5885365af78d7ae88588ef793bde26f0b8b33120
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>